feat: implemented signup functionality through context api

+++ src/context/AuthProvider.jsx
@@ -1,12 +1,45 @@
 import React, { useState } from "react";
 import { AuthContext } from "./AuthContext";
+import { createUser, getUsers } from "../services/userApi";
+import { Navigate } from "react-router-dom";
 
 const AuthProvider = ({ children }) => {
   const [user, setUser] = useState(null);
 
-  const login = (data) => {
+  const signup = async (data) => {
+    const users = await getUsers();
+
+    const exists = users.find(
+      (u) => u.email === data.email
+    );
+
+    if (exists) {
+      alert("User already exists");
+      throw new Error("User already exists");
+    }
+
+    const newUser = await createUser(data);
+
     localStorage.setItem("token", "dummy-token");
-    setUser(data.email);
+    setUser(newUser);
+  };
+
+   const login = async (data) => {
+    const users = await getUsers();
+
+    const exists = users.find(
+      (u) =>
+        u.email === data.email &&
+        u.password === data.password
+    );
+
+    if (!exists) {
+      alert("Invalid email or password");
+      return;
+    }
+
+    localStorage.setItem("token", "dummy-token");
+    setUser(exists);
   };
 
   const logout = () => {
@@ -14,7 +47,7 @@ const AuthProvider = ({ children }) => {
     setUser(null);
   };
   return (
-    <AuthContext.Provider value={{ user, login, logout }}>
+    <AuthContext.Provider value={{ user, login, logout, signup }}>
       {children}
     </AuthContext.Provider>
   );
